requirements.txt
django>=5.0
djangorestframework>=3.15
psycopg2-binary
djangorestframework-simplejwt
faiss-cpu
django-cors-headers
channels==4.1.0
daphne==4.1.0
pytest-django
python-dotenv
